requirements.txt
astroid==2.4.2
cffi==1.14.3
click==7.1.2
cryptography==3.2.1
Flask==1.1.2
isort==5.6.4
itsdangerous==1.1.0
Jinja2==2.11.2
lazy-object-proxy==1.4.3
MarkupSafe==1.1.1
mccabe==0.6.1
passlib==1.7.4
psycopg2-binary==2.8.6
pycparser==2.20
PyJWT==1.7.1
pylint==2.6.0
python-dotenv==0.15.0
six==1.15.0
toml==0.10.1
Werkzeug==1.0.1
wrapt==1.12.1
